+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 ### Enhancements
 * Support for new SchemaMode::HardResetFile added. ([#4782](https://github.com/realm/realm-core/issues/4782))
 * Support for keypaths in change notifications added to C-API ([#5216](https://github.com/realm/realm-core/issues/5216))
+* Release of callback functions done through realm_release() ([#5217](https://github.com/realm/realm-core/issues/5217))
 
 ### Fixed
 * Query parser would not accept "in" as a property name ([#5312](https://github.com/realm/realm-core/issues/5312))
@@ -36,6 +37,7 @@
  
 ### Breaking changes
 * SchemaMode::ResetFile renamed to SchemaMode::SoftResetFile.
+* Token type changed for registration of callback functions for changes on Realm and Schema. The functions are unregistered be releasing the token through 'realm_release()'.
 
 ### Compatibility
 * Fileformat: Generates files with format v22. Reads and automatically upgrade from fileformat v5.